+++ js/util.js
@@ -13,41 +13,10 @@ function getRandomMessage() {
   return result.join(' ');
 }
 
-// Функция: случайное имя
+// Функция: рандомное выбор случайного индекса из массива
 function getRandomName() {
-  return names[getRandomInt(0, names.length - 1)];
+  return Math.floor(Math.random()*arr.lenght);
 }
 
-// Функция: генерация массива комментариев
-function generateComments() {
-  const commentsCount = getRandomInt(0, 30); // 0–30 комментариев
-  const comments = [];
-  const usedIds = new Set();
-
-  for (let i = 0; i < commentsCount; i++) {
-    let id;
-    // Уникальный id
-    do {
-      id = getRandomInt(1, 100000);
-    } while (usedIds.has(id));
-    usedIds.add(id);
-
-    // Формируем объект комментария: name в конце
-    comments.push({
-      id: id,
-      // eslint-disable-next-line no-template-curly-in-string
-      avatar: 'img/avatar-${getRandomInt(1, 6)}.svg',
-      message: getRandomMessage(),
-      name: getRandomName()
-    });
-  }
-
-  return comments;
-}
-
-// Пример вызова
-comments = generateComments();
-//console.log(comments);
-
-export {getRandomInt,generateComments};
+export {getRandomInt,getRandomMessage,getRandomName};
